requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.1.3
django-autoslug==1.9.9
django-restframework==0.0.1
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-yasg==1.21.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
pillow==11.0.0
psycopg==3.2.3
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
